requirements_energy.txt
# Core ML Frameworks
torch>=2.0.0
torchvision>=0.15.0
jax[cuda12]>=0.4.0
flax>=0.7.0

# Energy and Hardware Monitoring (CRITICAL)
nvidia-ml-py>=12.535.0  # Modern NVML bindings (replaces deprecated pynvml)
psutil>=5.9.0   # System and process monitoring

# Data Handling
datasets>=2.13.0,<2.14.0  # Compatible with transformers<=4.57.3 (requires huggingface-hub<1.0)
transformers>=4.30.0  # Latest available version (4.57.3) with pre-built tokenizers wheels
huggingface-hub>=0.14.1,<1.0.0  # Explicit constraint compatible with both transformers and datasets
Pillow>=9.0.0
pyarrow>=14.0.0,<15.0.0  # Required by datasets' pyarrow backend

# Scientific Computing
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Experiment Tracking (Optional)
wandb>=0.15.0

# Utilities
tqdm>=4.65.0
